# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface JsonObject {
  [key: string]: unknown;
}

export type Right = (typeof Right)[keyof typeof Right];

export const Right = {
  manage_profiles: "manage_profiles",
  assign_roles: "assign_roles",
  manage_roles: "manage_roles",
  manage_org_units: "manage_org_units",
  create_checklists: "create_checklists",
  edit_checklists: "edit_checklists",
  delete_checklists: "delete_checklists",
  view_reports: "view_reports",
  manage_checklist_settings: "manage_checklist_settings",
} as const;

export type OrgUnitType = (typeof OrgUnitType)[keyof typeof OrgUnitType];

export const OrgUnitType = {
  region: "region",
  district: "district",
  location: "location",
} as const;

export interface OrgUnit {
  id: number;
  tenantId: number;
  parentId?: number | null;
  name: string;
  type: OrgUnitType;
  createdAt: string;
}

export interface CreateOrgUnitRequest {
  /** @minLength 1 */
  name: string;
  type: OrgUnitType;
  parentId?: number | null;
}

export interface UpdateOrgUnitRequest {
  /** @minLength 1 */
  name?: string;
  parentId?: number | null;
}

export interface Role {
  id: number;
  tenantId?: number | null;
  name: string;
  level: number;
  isSystem: boolean;
  rights: Right[];
}

export interface CreateRoleRequest {
  /** @minLength 1 */
  name: string;
  level: number;
  rights: Right[];
}

export interface UpdateRoleRequest {
  /** @minLength 1 */
  name?: string;
  level?: number;
  rights?: Right[];
}

export interface Profile {
  id: number;
  tenantId?: number | null;
  orgUnitId?: number | null;
  orgUnit?: OrgUnit | null;
  username: string;
  displayName: string;
  roleId: number;
  role: Role;
  mustChangePassword: boolean;
  isActive: boolean;
}

export interface CreateProfileRequest {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  displayName: string;
  /**
   * @minLength 12
   * @pattern ^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9]).{12,}$
   */
  password: string;
  roleId: number;
  orgUnitId?: number | null;
  mustChangePassword?: boolean;
}

export interface UpdateProfileRequest {
  /** @minLength 1 */
  username?: string;
  /** @minLength 1 */
  displayName?: string;
  /**
   * @minLength 12
   * @pattern ^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9]).{12,}$
   */
  password?: string;
  roleId?: number;
  orgUnitId?: number | null;
  isActive?: boolean;
  mustChangePassword?: boolean;
}

export interface UpdateMeRequest {
  /** @minLength 1 */
  displayName?: string;
  currentPassword?: string;
  /**
   * @minLength 12
   * @pattern ^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9]).{12,}$
   */
  newPassword?: string;
}

export type RegisterRequestBusinessType =
  (typeof RegisterRequestBusinessType)[keyof typeof RegisterRequestBusinessType];

export const RegisterRequestBusinessType = {
  "single-unit": "single-unit",
  "multi-unit": "multi-unit",
} as const;

export interface RegisterRequest {
  /** @minLength 1 */
  email: string;
  /**
   * @minLength 12
   * @pattern ^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9]).{12,}$
   */
  password: string;
  /** @minLength 1 */
  businessName: string;
  businessType: RegisterRequestBusinessType;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface RefreshRequest {
  refreshToken: string;
}

export interface LogoutRequest {
  refreshToken?: string;
}

export interface AuthResult {
  accessToken: string;
  refreshToken: string;
  profile: Profile;
}

export interface PasskeyCredential {
  id: number;
  credentialId: string;
  label?: string | null;
  createdAt: string;
  lastUsedAt?: string | null;
}

export interface PasskeyRegisterVerifyRequest {
  response: JsonObject;
  label?: string;
}

export interface PasskeyAuthOptionsRequest {
  username?: string;
}

export interface PasskeyAuthVerifyRequest {
  response: JsonObject;
  username?: string;
}

export type OnboardingCompleteRequestTeamMembersItem = {
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  role: string;
  orgUnitId?: number | null;
};

export interface OnboardingCompleteRequest {
  teamMembers: OnboardingCompleteRequestTeamMembersItem[];
}

export type OnboardingCompleteResponseFailedItem = {
  email: string;
  error: string;
};

export interface OnboardingCompleteResponse {
  invited: number;
  failed: OnboardingCompleteResponseFailedItem[];
}