# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with username + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Tokens + profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a new account (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created — tokens + profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "409":
          description: Email already in use
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "204":
          description: Signed out
  /auth/refresh:
    post:
      operationId: refresh
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /auth/passkey/register-verify:
    post:
      operationId: passkeyRegisterVerify
      tags: [auth]
      summary: Verify passkey registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyRegisterVerifyRequest"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyCredential"
  /auth/passkey/auth-options:
    post:
      operationId: passkeyAuthOptions
      tags: [auth]
      summary: Generate passkey assertion options
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyAuthOptionsRequest"
      responses:
        "200":
          description: Options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsonObject"
  /auth/passkey/auth-verify:
    post:
      operationId: passkeyAuthVerify
      tags: [auth]
      summary: Verify passkey assertion and sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyAuthVerifyRequest"
      responses:
        "200":
          description: Tokens + profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /org-units:
    get:
      operationId: listOrgUnits
      tags: [org-units]
      summary: List org units for the caller's tenant
      responses:
        "200":
          description: Org units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrgUnit"
    post:
      operationId: createOrgUnit
      tags: [org-units]
      summary: Create an org unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgUnitRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgUnit"
  /org-units/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      operationId: updateOrgUnit
      tags: [org-units]
      summary: Update an org unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgUnitRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgUnit"
    delete:
      operationId: deleteOrgUnit
      tags: [org-units]
      summary: Delete an org unit (must have no children and no assigned users)
      responses:
        "204":
          description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
    Right:
      type: string
      enum:
        - manage_profiles
        - assign_roles
        - manage_roles
        - manage_org_units
        - create_checklists
        - edit_checklists
        - delete_checklists
        - view_reports
        - manage_checklist_settings
    OrgUnitType:
      type: string
      enum: [region, district, location]
    OrgUnit:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        parentId:
          type: integer
          nullable: true
        name:
          type: string
        type:
          $ref: "#/components/schemas/OrgUnitType"
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, name, type, createdAt]
    CreateOrgUnitRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          $ref: "#/components/schemas/OrgUnitType"
        parentId:
          type: integer
          nullable: true
      required: [name, type]
    UpdateOrgUnitRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        parentId:
          type: integer
          nullable: true
    Role:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
          nullable: true
        name:
          type: string
        level:
          type: integer
        isSystem:
          type: boolean
        rights:
          type: array
          items:
            $ref: "#/components/schemas/Right"
      required: [id, name, level, isSystem, rights]
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
          nullable: true
        orgUnitId:
          type: integer
          nullable: true
        orgUnit:
          $ref: "#/components/schemas/OrgUnit"
          nullable: true
        username:
          type: string
        displayName:
          type: string
        roleId:
          type: integer
        role:
          $ref: "#/components/schemas/Role"
        mustChangePassword:
          type: boolean
        isActive:
          type: boolean
      required: [id, username, displayName, roleId, role, mustChangePassword, isActive]
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 12
          pattern: '^(?=.*[A-Z])(?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9]).{12,}$'
        businessName:
          type: string
          minLength: 1
        businessType:
          type: string
          enum: [single-unit, multi-unit]
      required: [email, password, businessName, businessType]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    LogoutRequest:
      type: object
      properties:
        refreshToken:
          type: string
    AuthResult:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        profile:
          $ref: "#/components/schemas/Profile"
      required: [accessToken, refreshToken, profile]
    PasskeyCredential:
      type: object
      properties:
        id:
          type: integer
        credentialId:
          type: string
        label:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, credentialId, createdAt]
    PasskeyRegisterVerifyRequest:
      type: object
      properties:
        response:
          $ref: "#/components/schemas/JsonObject"
        label:
          type: string
      required: [response]
    PasskeyAuthOptionsRequest:
      type: object
      properties:
        username:
          type: string
    PasskeyAuthVerifyRequest:
      type: object
      properties:
        response:
          $ref: "#/components/schemas/JsonObject"
        username:
          type: string
      required: [response]
...