# lib/api-zod/src/generated/types/createOrgUnitRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrgUnitType } from "./orgUnitType";

export interface CreateOrgUnitRequest {
  /** @minLength 1 */
  name: string;
  type: OrgUnitType;
  parentId?: number | null;
}