# lib/api-zod/src/generated/types/orgUnit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrgUnitType } from "./orgUnitType";

export interface OrgUnit {
  id: number;
  tenantId: number;
  parentId?: number | null;
  name: string;
  type: OrgUnitType;
  createdAt: Date;
}