# lib/api-zod/src/generated/types/orgUnitType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type OrgUnitType = (typeof OrgUnitType)[keyof typeof OrgUnitType];

export const OrgUnitType = {
  region: "region",
  district: "district",
  location: "location",
} as const;