# lib/api-zod/src/generated/types/right.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Right = (typeof Right)[keyof typeof Right];

export const Right = {
  manage_profiles: "manage_profiles",
  assign_roles: "assign_roles",
  manage_roles: "manage_roles",
  manage_org_units: "manage_org_units",
  create_checklists: "create_checklists",
  edit_checklists: "edit_checklists",
  delete_checklists: "delete_checklists",
  view_reports: "view_reports",
  manage_checklist_settings: "manage_checklist_settings",
} as const;