# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Right } from "./right";

export interface Role {
  id: number;
  tenantId?: number | null;
  name: string;
  level: number;
  isSystem: boolean;
  rights: Right[];
}