# lib/api-zod/src/generated/types/updateOrgUnitRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOrgUnitRequest {
  /** @minLength 1 */
  name?: string;
  parentId?: number | null;
}